[3] Feat : Error 프로토콜 제작, CodingKeys 추가
