+++ Expo1900/Expo1900/Model/Error.swift
@@ -1,0 +1,14 @@
+//
+//  Error.swift
+//  Expo1900
+//
+//  Created by 천수현 on 2021/04/06.
+//
+
+import Foundation
+
+enum JsonDecodingError: Error {
+    case dataCorrupted(context: String)
+    case keyNotFound(key: String, context: String)
+    case valueNotFound(value: String, context: String)
+}
